Split the phone-index tests out of the sms webhook suite
The webhook test file carried two unrelated groups: how a phone number is
turned into a lookup value, and how that value is stored and found. Those move
to test/shared/sms/phone-index.test.ts, next to the code they cover, dropping
the webhook file from 478 to 380 lines.

Both files needed the same booked attendee with a phone number, so that setup
is now createTestAttendeeWithPhone in the shared attendee test helpers.
